pyproject.toml
[project]
name = "autoresearcher-litter-detection"
version = "0.1.0"
description = "Autonomous CNN training for litter segmentation"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "mlflow>=2.13.0",
    "datasets>=2.19.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.5.0",
    "pycocotools>=2.0.8",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
    "huggingface-hub>=0.23.0",
    "albumentations>=1.4.0",
    "opencv-python-headless>=4.10.0",
    "pydantic>=2.12.5",
    "eclipse-zenoh>=1.9.0",
    "pydantic-settings>=2.13.1",
    "imutils>=0.5.4",
    "msgspec>=0.21.0",
    "loguru>=0.7.3",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "pandas-stubs~=2.3.3",
    "onnx>=1.16.0",
    "onnxruntime>=1.18.0",
    "onnxscript>=0.1.0",
    "nvidia-ml-py>=12.0.0",
    "segmentation-models-pytorch>=0.5.0",
    "opencv-python>=4.13.0.92",
    "pydantic-ai>=0.0.20",
]

[project.scripts]
camera = "litter_detector.camera.main:main"
detector = "litter_detector.detector.main:main"
mission = "litter_detector.agents.main:main"

[tool.uv]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/litter_detector"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
    "ty>=0.0.29",
]
